+++ src/main/java/pluralsight/respository/HibernateSpeakerRepositoryImpl.java
@@ -1,13 +1,14 @@
-package respository;
+package pluralsight.respository;
 
-import model.Speaker;
+import pluralsight.model.Speaker;
+import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
 
+@Repository("speakerRepository")
 public class HibernateSpeakerRepositoryImpl implements SpeakerRepository {
 
-    @Override
     public List<Speaker> findAll(){
 
         List<Speaker> speakers = new ArrayList<>();
